Rob's Test Checkin
This is a test of what it looks like when someone makes a branch off of the "Master" branch, makes changes, commits the changes and then makes a "Pull Request" that needs to be "merged" back into master.  You should never work in the master branch... you should make a branch for each feature or one with your name or whatever and work in there.  When you're done, you check it in and make a pull request that you can merge yourself.